factor out Fixtures and HTTP helpers
