+++ e2e/fixtures.go
@@ -1,0 +1,141 @@
+// Copyright Project Contour Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build e2e
+
+package e2e
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+// Fixtures holds references to all of the E2E fixtures helpers.
+type Fixtures struct {
+	// Echo provides helpers for working with the ingress-conformance-echo
+	// test fixture.
+	Echo *Echo
+}
+
+// Echo manages the ingress-conformance-echo fixture.
+type Echo struct {
+	client client.Client
+	t      *testing.T
+}
+
+// Create creates the ingress-conformance-echo fixture, specifically
+// the deployment and service, in the given namespace and with the given name, or
+// fails the test if it encounters an error. Namespace is defaulted to "default"
+// and name is defaulted to "ingress-conformance-echo" if not provided.
+func (e *Echo) Create(ns, name string) {
+	valOrDefault := func(val, defaultVal string) string {
+		if val != "" {
+			return val
+		}
+		return defaultVal
+	}
+
+	ns = valOrDefault(ns, "default")
+	name = valOrDefault(name, "ingress-conformance-echo")
+
+	deployment := &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: ns,
+			Name:      name,
+		},
+		Spec: appsv1.DeploymentSpec{
+			Selector: &metav1.LabelSelector{
+				MatchLabels: map[string]string{"app.kubernetes.io/name": name},
+			},
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels: map[string]string{"app.kubernetes.io/name": name},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name:  "conformance-echo",
+							Image: "k8s.gcr.io/ingressconformance/echoserver:v0.0.1",
+							Env: []corev1.EnvVar{
+								{
+									Name:  "INGRESS_NAME",
+									Value: name,
+								},
+								{
+									Name:  "SERVICE_NAME",
+									Value: name,
+								},
+								{
+									Name: "POD_NAME",
+									ValueFrom: &corev1.EnvVarSource{
+										FieldRef: &corev1.ObjectFieldSelector{
+											FieldPath: "metadata.name",
+										},
+									},
+								},
+								{
+									Name: "NAMESPACE",
+									ValueFrom: &corev1.EnvVarSource{
+										FieldRef: &corev1.ObjectFieldSelector{
+											FieldPath: "metadata.namespace",
+										},
+									},
+								},
+							},
+							Ports: []corev1.ContainerPort{
+								{
+									Name:          "http-api",
+									ContainerPort: 3000,
+								},
+							},
+							ReadinessProbe: &corev1.Probe{
+								Handler: corev1.Handler{
+									HTTPGet: &corev1.HTTPGetAction{
+										Path: "/health",
+										Port: intstr.FromInt(3000),
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+	require.NoError(e.t, e.client.Create(context.TODO(), deployment))
+
+	service := &corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: ns,
+			Name:      name,
+		},
+		Spec: corev1.ServiceSpec{
+			Ports: []corev1.ServicePort{
+				{
+					Name:       "http",
+					Port:       80,
+					TargetPort: intstr.FromString("http-api"),
+				},
+			},
+			Selector: map[string]string{"app.kubernetes.io/name": name},
+		},
+	}
+	require.NoError(e.t, e.client.Create(context.TODO(), service))
+}
