+++ e2e/gateway/002_header_condition_match_test.go
@@ -33,7 +33,7 @@ func testGatewayHeaderConditionMatch(t *testing.T, fx *e2e.Framework) {
 	fx.CreateNamespace(namespace)
 	defer fx.DeleteNamespace(namespace)
 
-	fx.CreateEchoWorkload(namespace, "echo-header-exact")
+	fx.Fixtures.Echo.Create(namespace, "echo-header-exact")
 
 	// HTTPRoute
 	route := &gatewayv1alpha1.HTTPRoute{
@@ -97,7 +97,7 @@ func testGatewayHeaderConditionMatch(t *testing.T, fx *e2e.Framework) {
 	}
 
 	for _, tc := range cases {
-		res, ok := fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+		res, ok := fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 			Host: string(route.Spec.Hostnames[0]),
 			RequestOpts: []func(*http.Request){
 				e2e.OptSetHeaders(tc.headers),
