+++ e2e/httpproxy/006_merge_slash_test.go
@@ -31,7 +31,7 @@ func testMergeSlash(t *testing.T, fx *e2e.Framework) {
 	fx.CreateNamespace(namespace)
 	defer fx.DeleteNamespace(namespace)
 
-	fx.CreateEchoWorkload(namespace, "ingress-conformance-echo")
+	fx.Fixtures.Echo.Create(namespace, "ingress-conformance-echo")
 
 	p := &contourv1.HTTPProxy{
 		ObjectMeta: metav1.ObjectMeta{
@@ -61,7 +61,7 @@ func testMergeSlash(t *testing.T, fx *e2e.Framework) {
 	}
 	fx.CreateHTTPProxyAndWaitFor(p, httpProxyValid)
 
-	res, ok := fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+	res, ok := fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 		Host:      p.Spec.VirtualHost.Fqdn,
 		Path:      "/anything/this//has//lots////of/slashes",
 		Condition: e2e.HasStatusCode(200),
