+++ e2e/httpproxy/019_local_rate_limiting_test.go
@@ -31,7 +31,7 @@ func testLocalRateLimitingVirtualHost(t *testing.T, fx *e2e.Framework) {
 	fx.CreateNamespace(namespace)
 	defer fx.DeleteNamespace(namespace)
 
-	fx.CreateEchoWorkload(namespace, "echo")
+	fx.Fixtures.Echo.Create(namespace, "echo")
 
 	p := &contourv1.HTTPProxy{
 		ObjectMeta: metav1.ObjectMeta{
@@ -58,7 +58,7 @@ func testLocalRateLimitingVirtualHost(t *testing.T, fx *e2e.Framework) {
 
 	// Wait until we get a 200 from the proxy confirming
 	// the pods are up and serving traffic.
-	_, ok := fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+	_, ok := fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 		Host:      p.Spec.VirtualHost.Fqdn,
 		Condition: e2e.HasStatusCode(200),
 	})
@@ -77,7 +77,7 @@ func testLocalRateLimitingVirtualHost(t *testing.T, fx *e2e.Framework) {
 	// is returned since we're allowed one request per hour.
 	//
 	// TODO it'd be better to just make a single request.
-	_, ok = fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+	_, ok = fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 		Host:      p.Spec.VirtualHost.Fqdn,
 		Condition: e2e.HasStatusCode(200),
 	})
@@ -87,7 +87,7 @@ func testLocalRateLimitingVirtualHost(t *testing.T, fx *e2e.Framework) {
 	// is now gotten since we've exceeded the rate limit.
 	//
 	// TODO it'd be better to just make a single request.
-	_, ok = fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+	_, ok = fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 		Host:      p.Spec.VirtualHost.Fqdn,
 		Condition: e2e.HasStatusCode(429),
 	})
@@ -100,7 +100,7 @@ func testLocalRateLimitingRoute(t *testing.T, fx *e2e.Framework) {
 	fx.CreateNamespace(namespace)
 	defer fx.DeleteNamespace(namespace)
 
-	fx.CreateEchoWorkload(namespace, "echo")
+	fx.Fixtures.Echo.Create(namespace, "echo")
 
 	p := &contourv1.HTTPProxy{
 		ObjectMeta: metav1.ObjectMeta{
@@ -140,7 +140,7 @@ func testLocalRateLimitingRoute(t *testing.T, fx *e2e.Framework) {
 
 	// Wait until we get a 200 from the proxy confirming
 	// the pods are up and serving traffic.
-	_, ok := fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+	_, ok := fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 		Host:      p.Spec.VirtualHost.Fqdn,
 		Condition: e2e.HasStatusCode(200),
 	})
@@ -159,7 +159,7 @@ func testLocalRateLimitingRoute(t *testing.T, fx *e2e.Framework) {
 	// is returned since we're allowed one request per hour.
 	//
 	// TODO it'd be better to just make a single request.
-	_, ok = fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+	_, ok = fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 		Host:      p.Spec.VirtualHost.Fqdn,
 		Condition: e2e.HasStatusCode(200),
 	})
@@ -169,7 +169,7 @@ func testLocalRateLimitingRoute(t *testing.T, fx *e2e.Framework) {
 	// is now gotten since we've exceeded the rate limit.
 	//
 	// TODO it'd be better to just make a single request.
-	_, ok = fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+	_, ok = fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 		Host:      p.Spec.VirtualHost.Fqdn,
 		Condition: e2e.HasStatusCode(429),
 	})
@@ -177,7 +177,7 @@ func testLocalRateLimitingRoute(t *testing.T, fx *e2e.Framework) {
 
 	// Make a request against the route that doesn't have rate limiting
 	// to confirm we still get a 200 for that route.
-	_, ok = fx.HTTPRequestUntil(&e2e.HTTPRequestOpts{
+	_, ok = fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
 		Host:      p.Spec.VirtualHost.Fqdn,
 		Path:      "/unlimited",
 		Condition: e2e.HasStatusCode(200),
